DDCE-5971: Pbik-frontend - PBIK - Scala 3 Upgrade (#304)
DDCE-5971: Pbik-frontend - PBIK - Scala 3 Upgrade

+++ app/repositories/DefaultSessionRepository.scala
@@ -43,7 +43,7 @@ class DefaultSessionRepository @Inject() (config: PbikAppConfig, mongo: MongoCom
           ascending("lastUpdated"),
           IndexOptions()
             .name("userAnswersExpiry")
-            .expireAfter(config.mongoTTL, TimeUnit.SECONDS)
+            .expireAfter(config.mongoTTL.toLong, TimeUnit.SECONDS)
         )
       )
     )
